+++ sheets/EnhancedJournalSheet.js
@@ -1082,4 +1082,32 @@ export class EnhancedJournalSheet extends JournalSheet {
 
         return items;
     }
+
+    async _onItemSummary(event) {
+        event.preventDefault();
+
+        let li = $(event.currentTarget).closest('li.item');
+
+        const id = li.data("id");
+        let itemData = (this.object.getFlag('monks-enhanced-journal', 'items') || []).find(i => i._id == id);
+        if (!itemData)
+            return;
+
+        let item = new CONFIG.Item.documentClass(itemData);
+        const chatData = item.getChatData({ secrets: false });
+
+        // Toggle summary
+        if (li.hasClass("expanded")) {
+            let summary = li.children(".item-summary");
+            summary.slideUp(200, () => summary.remove());
+        } else {
+            let div = $(`<div class="item-summary">${chatData.description.value}</div>`);
+            let props = $('<div class="item-properties"></div>');
+            chatData.properties.forEach(p => props.append(`<span class="tag">${p}</span>`));
+            div.append(props);
+            li.append(div.hide());
+            div.slideDown(200);
+        }
+        li.toggleClass("expanded");
+    }
 }